requirements.txt
pyyaml
requests
bs4
requests_cache
tabulate
pyfiglet
